refactor store helpers to separate module, fix test
